Make print_navigator more clear. Make some doubles to floats, as double precision is not needed. Change a comment. Move min_fall_height_for_sound to movement.c. Make INIT_BEGIN_LEVEL_FADE to BEGIN_LEVEL_FADE. Make DEBUG_BYTE have brackets around it, so it works in if statements with no brackets.

+++ src/main/global_types.h
@@ -93,8 +93,8 @@ typedef struct {
 typedef struct {
 	vec3D pos; // x, y, z
 	const vec3D dir;
-	double dist;
-	const double step;
+	float dist; // 32-bit dist used here b/c pos and dir have float components
+	const float step;
 } Tracer;
 
 //////////
